+++ src/main/java/controller/ForwardController.java
@@ -1,8 +1,8 @@
 package controller;
 
 import utils.FileIoUtils;
-import webserver.HttpRequest;
-import webserver.HttpResponse;
+import domain.HttpRequest;
+import domain.HttpResponse;
 
 import java.io.IOException;
 import java.net.URISyntaxException;
