+++ src/main/java/domain/Cookie.java
@@ -1,6 +1,4 @@
-package model;
-
-import webserver.HttpRequest;
+package domain;
 
 import java.util.HashMap;
 import java.util.Map;
