+++ src/main/java/domain/HttpRequest.java
@@ -1,4 +1,4 @@
-package webserver;
+package domain;
 
 import java.io.BufferedReader;
 import java.io.IOException;
