+++ src/main/java/domain/HttpResponse.java
@@ -1,7 +1,8 @@
-package webserver;
+package domain;
 
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
+import webserver.RequestHandler;
 
 import java.io.DataOutputStream;
 import java.io.IOException;
