+++ src/test/java/webserver/ControllerTest.java
@@ -4,17 +4,17 @@ import controller.UserCreateController;
 import controller.UserListController;
 import controller.UserLoginController;
 import db.DataBase;
+import domain.HttpRequest;
+import domain.HttpResponse;
 import model.User;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 
-import javax.xml.crypto.Data;
 import java.io.*;
 import java.net.URISyntaxException;
 import java.nio.file.Files;
 import java.nio.file.Path;
 import java.nio.file.Paths;
-import java.util.Arrays;
 
 import static org.assertj.core.api.Assertions.assertThat;
